Organize dart screen files similar to user paths
