+++ lib/screens/home/settings/settings.dart
@@ -1,6 +1,6 @@
 import 'package:bitcoin_ui/bitcoin_ui.dart';
 import 'package:donationwallet/global_functions.dart';
-import 'package:donationwallet/home.dart';
+import 'package:donationwallet/screens/home/home.dart';
 import 'package:donationwallet/generated/rust/api/wallet.dart';
 import 'package:donationwallet/services/settings_service.dart';
 import 'package:donationwallet/states/chain_state.dart';
