+++ lib/screens/home/wallet/wallet.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:bitcoin_ui/bitcoin_ui.dart';
 import 'package:donationwallet/global_functions.dart';
 import 'package:donationwallet/services/synchronization_service.dart';
-import 'package:donationwallet/spend.dart';
+import 'package:donationwallet/screens/home/wallet/spend/spend.dart';
 import 'package:donationwallet/states/chain_state.dart';
 import 'package:donationwallet/states/wallet_state.dart';
 import 'package:flutter/material.dart';
